replaced numpy.flatten with ivy.reshape
